library getResources issue fixed and img format changes
